requirements.txt
Flask
python-dotenv
google-cloud-logging
google-cloud-firestore
google-cloud-tasks
requests
gunicorn